Bookstore: Use Redux in React components.
In this milestone I  Wrap my <App> with the <Provider> component from react-redux. Passed configured store into the <Provider>. Created an initialState variable for booksSlice. Display books, received from the slice, in a reusable component. Dispatch actions using useDispatch. Added an event handler that adds a book to the books array. Added an event handler that removes a book from the books array